+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1258
-    const val appVersion = "4.21.10.3"
+    const val appCodeNumber = 1259
+    const val appVersion = "4.21.10.4"
     const val identifier = "com.unciv.app"
 }